hide unpack_stack in source file
